Move RacesList into seperate class
